Found 4 bugs in atomic configuration
1. All values were being read from the json file as strings.
   We even had a value whose absence was labeled "clear".
   Even though Json has data types and freaking null.
2. Userspace jool in atomic config mode was sending all numbers as
   if they were __u64s but the kernel module was expecting more
   specific sizes.
3. These __u64 values were creating unwanted slop in the byte
   stream.

Bugs 2 and 3 sort of canceled each other in my setup, which is why
I hadn't seen them earlier.

I haven't yet fixed the fourth bug:

4. Userspace jool is forgetting to add a header to subsequent
   atomic config packets, which leads to chunks of configuration
   not committing.

+++ mod/stateful/joold.c
@@ -196,11 +196,7 @@ static bool should_send(struct joold_queue *queue)
 	if (queue->count == 0)
 		return false;
 
-	/*
-	 * This cast is fair; incoming validations prevent this number
-	 * from being higher than 0xFFFFFFFF even on 64-bit machines.
-	 */
-	deadline = (unsigned long)queue->config.flush_deadline;
+	deadline = queue->config.flush_deadline;
 	if (time_before(queue->last_flush_time + deadline, jiffies))
 		return true;
